I did some stuff
because we now have gRPC clients and HTTP clients, I've moved many
of the old 'client' files to 'http_client' paths.

I've also implemented a crude first-pass of an ItemsService (only)
via gRPC, as well as a simplistic integration test. That integration
test fails at the moment (appropriately), but that's because I need to
implement auth for it, which I will do next.